Refactor memory and byte order handling, improve PRNG interface, and enhance string utilities
- Updated BinaryWriter to use UINT8 instead of PVOID for address handling and implemented placement new/delete operators.
- Refactored ByteOrder macros into a static class with constexpr methods for byte order swapping.
- Changed PRNG GetArray method to return VOID instead of INT32, simplifying the interface.
- Enhanced memory operations in memory.cc to handle aligned memory copying and added bzero implementation.
- Improved string utilities for better type safety and null pointer handling in comparison functions.
- Updated network socket handling to use the new ByteOrder class for consistency in byte order operations.
- Refactored TLS handling to utilize the ByteOrder class for all byte order conversions.
- Adjusted tests to reflect changes in PRNG and memory operations, ensuring they validate expected behavior without relying on return values.

+++ src/core/compiler/compiler_runtime.i386.cc
@@ -163,10 +163,11 @@ extern "C"
 		}
 
 		// Determine result signs and convert to absolute values
+		// Cast to UINT64 before negating to avoid signed overflow UB on INT64_MIN
 		const BOOL negNum = numerator < 0;
 		const BOOL negQuot = negNum != (denominator < 0);
-		const UINT64 absNum = (UINT64)(negNum ? -numerator : numerator);
-		const UINT64 absDen = (UINT64)(denominator < 0 ? -denominator : denominator);
+		const UINT64 absNum = negNum ? -(UINT64)numerator : (UINT64)numerator;
+		const UINT64 absDen = (denominator < 0) ? -(UINT64)denominator : (UINT64)denominator;
 
 		// Perform unsigned division on absolute values
 		UINT64 q, r;
